# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: AI 智能下单与生产协同平台 API
  version: 1.0.0
  description: |
    牙科定制工厂三端系统（医生端 / 客服端 / 生产端 / 管理端）后端 API 规范。

    ## 认证方式
    所有接口（除登录外）必须携带 Authorization: Bearer {access_token}。
    服务端对每个接口强制校验角色权限，禁止仅靠前端隐藏。

    ## 角色枚举
    - DOCTOR：医生
    - CS：客服
    - WORKER：技工/生产人员
    - ADMIN：超级管理员

    ## 通用响应格式
    所有接口统一返回：
    { "code": 200, "msg": "success", "data": {} }
    错误时：
    { "code": 4xx/5xx, "msg": "错误说明", "data": null }
  license:
    name: Proprietary internal API
    url: https://yourdomain.com/legal/proprietary-api-license
servers:
- url: https://api.yourdomain.com/v1
  description: 正式环境
- url: https://test-api.yourdomain.com/v1
  description: 测试环境
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    CommonResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        msg:
          type: string
          example: success
        data: {}
...
    ProductCatalogItem:
      type: object
      properties:
        product_id:
          type: integer
        product_type:
          type: string
          description: 产品类型编码，服务端统一转为大写下划线格式。
        product_name:
          type: string
        material_spec:
          type: string
          nullable: true
          description: 一期人工维护的材料/规格说明。
        base_price_cents:
          type: integer
          description: 一期人工维护基础价，单位分；不代表自动报价或最终结算。
        currency:
          type: string
          example: CNY
        status:
          type: string
          enum:
          - ACTIVE
          - INACTIVE
        price_note:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ProductCatalogRequest:
      type: object
      required:
      - product_type
      - product_name
      - base_price_cents
      properties:
        product_type:
          type: string
        product_name:
          type: string
        material_spec:
          type: string
        base_price_cents:
          type: integer
        currency:
          type: string
          default: CNY
        status:
          type: string
          enum:
          - ACTIVE
          - INACTIVE
        price_note:
          type: string
    UpdateProductCatalogRequest:
      type: object
      required:
      - product_name
      - base_price_cents
      properties:
        product_name:
          type: string
        material_spec:
          type: string
        base_price_cents:
          type: integer
        currency:
          type: string
          default: CNY
        status:
          type: string
          enum:
          - ACTIVE
          - INACTIVE
        price_note:
          type: string
...
    StaffWorkloadResponse:
      type: object
      description: |
        任务 9D.86 第一增量：人员档案 / 工作量看板只读响应。复用 system_user、system_dept、system_post、system_user_post、work_log 和 rework_record；不返回 password_hash、token、真实薪酬或工资结算字段。
      properties:
        user_id:
          type: integer
          format: int64
        username:
          type: string
        display_name:
          type: string
        user_type:
          type: string
          enum:
          - ADMIN
          - CS
          - WORKER
        status:
          type: string
        dept_id:
          type: integer
          format: int64
          nullable: true
        dept_name:
          type: string
          nullable: true
        post_names:
          type: array
          items:
            type: string
        role_codes:
          type: array
          items:
            type: string
        assigned_node_count:
          type: integer
          format: int64
          description: 已分配给该员工的工序节点数。
        active_node_count:
          type: integer
          format: int64
          description: READY / IN_PROGRESS 节点数。
        completed_work_log_count:
          type: integer
          format: int64
          description: 已完成工时记录数。
        effective_duration:
          type: integer
          format: int64
          description: 已完成工时有效时长，单位分钟。
        rework_count:
          type: integer
          format: int64
          description: 目标节点归属该员工的返工记录数。
        last_work_finished_at:
          type: string
          format: date-time
          nullable: true
        updated_at:
          type: string
          format: date-time
    StaffWorkloadListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            "$ref": "#/components/schemas/StaffWorkloadResponse"
        total:
          type: integer
          format: int64
        page:
          type: integer
        size:
          type: integer
...
    QualityRecordResponse:
      type: object
      description: |
        任务 9D.87 第一增量：质量记录只读响应。当前复用 check_record + rework_record 表达外返质量记录；
        不返回医生端脱敏外的内部生产敏感扩展，也不代表完整 quality_record 独立事实表已完成。
      properties:
        quality_record_id:
          type: integer
          format: int64
          description: 当前第一增量复用 check_id 作为质量记录 ID。
        quality_record_type:
          type: string
          enum:
          - EXTERNAL_RETURN
        order_id:
          type: integer
          format: int64
        order_no:
          type: string
        product_type:
          type: string
        clinic_name:
          type: string
        check_id:
          type: integer
          format: int64
        check_result:
          type: string
          enum:
          - PASS
          - FAIL
        rework_id:
          type: integer
          format: int64
          nullable: true
        reason_category:
          type: string
          nullable: true
        reason_detail:
          type: string
          nullable: true
        responsibility_type:
          type: string
          nullable: true
        status:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
          nullable: true
    QualityRecordListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            "$ref": "#/components/schemas/QualityRecordResponse"
        total:
          type: integer
          format: int64
        page:
          type: integer
        size:
          type: integer
    ExternalReturnQualityRecordRequest:
      type: object
      required:
      - order_id
      - reason_category
      - responsibility_type
      - reason_detail
      properties:
        order_id:
          type: integer
          format: int64
        reason_category:
          type: string
          description: 必须使用 /reworks/dictionaries 返回的 REASON_CATEGORY code。
        responsibility_type:
          type: string
          description: 必须使用 /reworks/dictionaries 返回的 RESPONSIBILITY_TYPE code。
        reason_detail:
          type: string
          maxLength: 512
...
    ErrorResponse:
      type: object
      required:
      - code
      - msg
      - data
      properties:
        code:
          type: integer
          example: 403
        msg:
          type: string
          example: forbidden
        data:
          nullable: true
          type: object
...
  responses:
    BadRequest:
      description: 请求参数错误或校验失败
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: 未登录或 Token 无效
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/ErrorResponse"
    Forbidden:
      description: 无权限或数据范围拒绝
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/ErrorResponse"
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/ErrorResponse"
    Conflict:
      description: 业务状态冲突或幂等冲突
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/ErrorResponse"
    ServiceUnavailable:
      description: 外部依赖或存储服务不可用
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/ErrorResponse"
    DefaultError:
      description: 未分类错误响应
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/ErrorResponse"
security:
- BearerAuth: []
paths:
...
  "/products":
    get:
      tags:
      - Product
      summary: 查询产品参数与基础价
      description: |
        权限：CS / ADMIN。
        任务 9D.90 第一增量：客服 / 管理员维护一期产品参数、材料规格和人工基础价；不做自动报价、客户分层价或最终结算。
      parameters:
      - name: keyword
        in: query
        schema:
          type: string
      - name: status
        in: query
        schema:
          type: string
          enum:
          - ACTIVE
          - INACTIVE
      - name: page
        in: query
        schema:
          type: integer
          default: 1
      - name: size
        in: query
        schema:
          type: integer
          default: 20
      responses:
        '200':
          description: 产品目录分页列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      items:
                        type: array
                        items:
                          "$ref": "#/components/schemas/ProductCatalogItem"
                      total:
                        type: integer
                      page:
                        type: integer
                      size:
                        type: integer
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '409':
          "$ref": "#/components/responses/Conflict"
        default:
          "$ref": "#/components/responses/DefaultError"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
      operationId: getProducts
    post:
      tags:
      - Product
      summary: 新增产品参数与基础价
      description: 权限：CS / ADMIN；基础价为人工维护口径，不触发自动报价或账单重算。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/ProductCatalogRequest"
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    "$ref": "#/components/schemas/ProductCatalogItem"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '409':
          "$ref": "#/components/responses/Conflict"
        default:
          "$ref": "#/components/responses/DefaultError"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
      operationId: postProducts
  "/products/{productId}":
    put:
      tags:
      - Product
      summary: 编辑产品参数与基础价
      description: 权限：CS / ADMIN；支持更新产品名称、材料规格、基础价、币种、状态和价格备注。
      parameters:
      - name: productId
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/UpdateProductCatalogRequest"
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    "$ref": "#/components/schemas/ProductCatalogItem"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '409':
          "$ref": "#/components/responses/Conflict"
        default:
          "$ref": "#/components/responses/DefaultError"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
      operationId: putProductsByProductId
...
  "/quality-records":
    get:
      tags:
      - Quality
      summary: 查询质量记录
      description: |
        任务 9D.87 第一增量：ADMIN / CS 查询内部质量记录。
        当前仅覆盖 EXTERNAL_RETURN 外返质量记录，底层复用 check_record 与 rework_record；医生端禁止读取。
      parameters:
      - name: record_type
        in: query
        required: false
        schema:
          type: string
          enum:
          - EXTERNAL_RETURN
      - name: status
        in: query
        required: false
        schema:
          type: string
      - name: responsibility_type
        in: query
        required: false
        schema:
          type: string
      - name: order_id
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: page
        in: query
        required: false
        schema:
          type: integer
          default: 1
      - name: size
        in: query
        required: false
        schema:
          type: integer
          default: 20
      responses:
        '200':
          description: 质量记录列表
          content:
            application/json:
              schema:
                allOf:
                - "$ref": "#/components/schemas/CommonResponse"
                - type: object
                  properties:
                    data:
                      "$ref": "#/components/schemas/QualityRecordListResponse"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '409':
          "$ref": "#/components/responses/Conflict"
        default:
          "$ref": "#/components/responses/DefaultError"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
      operationId: listQualityRecords
  "/quality-records/external-returns":
    post:
      tags:
      - Quality
      summary: 登记外返质量记录
      description: |
        任务 9D.87 第一增量：ADMIN / CS 登记外返质量记录。
        创建 check_type=EXTERNAL_RETURN 的 check_record，并创建 PENDING rework_record；不接真实退货 / 投诉系统。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/ExternalReturnQualityRecordRequest"
      responses:
        '200':
          description: 已登记外返质量记录
          content:
            application/json:
              schema:
                allOf:
                - "$ref": "#/components/schemas/CommonResponse"
                - type: object
                  properties:
                    data:
                      "$ref": "#/components/schemas/QualityRecordResponse"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '409':
          "$ref": "#/components/responses/Conflict"
        default:
          "$ref": "#/components/responses/DefaultError"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
      operationId: createExternalReturnQualityRecord
...
  "/staff/workload":
    get:
      tags:
      - Staff
      summary: 获取人员档案与工作量看板
      description: |
        任务 9D.86 第一增量：按最新版 PRD V2.0 补人员管理一期只读闭环。
        权限：ADMIN / CS 可查看内部员工；WORKER 只返回本人；DOCTOR 禁止访问。
        响应复用既有 RuoYi 风格用户、部门、岗位、角色和工时 / 返工事实，不返回 password_hash、token、薪酬、工资结算或客户隐私。
      parameters:
      - name: keyword
        in: query
        description: 按员工账号、姓名、部门或岗位模糊搜索。
        schema:
          type: string
      - name: page
        in: query
        schema:
          type: integer
          default: 1
      - name: size
        in: query
        schema:
          type: integer
          default: 20
          maximum: 100
      responses:
        '200':
          description: 人员档案与工作量列表
          content:
            application/json:
              schema:
                allOf:
                - "$ref": "#/components/schemas/CommonResponse"
                - type: object
                  properties:
                    data:
                      "$ref": "#/components/schemas/StaffWorkloadListResponse"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '409':
          "$ref": "#/components/responses/Conflict"
        default:
          "$ref": "#/components/responses/DefaultError"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
      operationId: getStaffWorkload
...